Move more tests out of `electron-browser` (#175356)
* Move more tests out of `electron-browser`

* use a safer layer for this one
