+++ src/vs/workbench/contrib/extensions/test/electron-sandbox/extension.test.ts
